optimzed color variables, added static example

+++ NeoPixelEffects.h
@@ -50,9 +50,9 @@ enum Effect {
 };
 
 struct EffectColor {
-  int r;
-  int g;
-  int b;
+  uint8_t r;
+  uint8_t g;
+  uint8_t b;
 };
 
 class NeoPixelEffects {
